refactor: moved mapping from entity to remapper, dispose controller and variable naming convention improved
