+++ lib/features/stream/data/remapper/todo_entity_convertor.dart
@@ -1,0 +1,18 @@
+import '../../domain/entities/todo_entities.dart';
+import '../models/todo_list_model.dart';
+
+class TodoEntityConvertor {
+
+  static List<TodoEntities> convertToTodoEntities(TodoListModel todoListModel) {
+    return todoListModel.todos!
+        .map(
+          (e) => TodoEntities(
+        id: e.id ?? 0,
+        title: e.todo ?? '',
+        completed: e.completed ?? false,
+      ),
+    )
+        .toList() ?? [];
+  }
+
+}
