+++ lib/features/stream/data/todo_repository_impl.dart
@@ -1,5 +1,6 @@
 import 'package:dio/dio.dart';
-import 'package:practice_stream/core/network/rest_clint.dart';
+import 'package:practice_stream/core/network/rest_client.dart';
+import 'package:practice_stream/features/stream/data/remapper/todo_entity_convertor.dart';
 import 'package:practice_stream/features/stream/domain/repository/todo_repository.dart';
 
 import '../domain/entities/todo_entities.dart';
@@ -16,6 +17,6 @@ class TodoRepositoryImpl extends TodoRepository {
   @override
   Future<List<TodoEntities>> getTodos() async {
     final response = await _restClient.getTodos();
-    return TodoEntities.convertToTodoEntities(response);
+    return TodoEntityConvertor.convertToTodoEntities(response);
   }
 }
