--- lib/features/stream/presentation/provider/todo_stream_provider.dart
@@ -1,19 +1,0 @@
-// import 'package:practice_stream/features/stream/domain/repository/todo_repository.dart';
-// import 'package:riverpod/riverpod.dart';
-// import 'package:riverpod_annotation/riverpod_annotation.dart';
-//
-// import '../../domain/entities/todo_entities.dart';
-// part 'todo_stream_provider.g.dart';
-//
-// @riverpod
-// class TodoStreamProvider extends _$TodoStreamProvider {
-//   late TodoRepository _todoRepository;
-//   @override
-//   List<TodoEntities>? build()
-//   {
-//     _todoRepository = ref.watch(todoRepositoryProvider);
-//     return [];
-//   }
-//
-//
-// }
