+++ pubspec.yaml
@@ -35,18 +35,20 @@ dependencies:
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^1.0.8
   riverpod: ^2.6.1
-  riverpod_annotation: ^2.6.1
-  build_runner: ^2.4.15
   dio: ^5.8.0+1
   retrofit: ^4.6.0
-  retrofit_generator: ^9.6.0
+  riverpod_annotation: ^2.6.1
   json_annotation: ^4.9.0
-  json_serializable: ^6.9.5
   rxdart: ^0.28.0
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
+  retrofit_generator: ^9.6.0
+  build_runner: ^2.4.15
+  json_serializable: ^6.9.5
+  riverpod_generator: ^2.6.1
+
 
   # The "flutter_lints" package below contains a set of recommended lints to
   # encourage good coding practices. The lint set provided by the package is
